starfield: refactor out
extra features/additions should reside in own files.
